+++ IbFlexReader/IbFlexReader.Contracts/Ib/TradeConfirm.cs
@@ -42,7 +42,7 @@
 
         public string Issuer { get; set; }
 
-        public int? Multiplier { get; set; }
+        public double? Multiplier { get; set; }
 
         public double? Strike { get; set; }
